// data/github/current/ui/packages/billing-app/components/pickers/__generated__/PaginatedRepositoryPickerRepos.graphql.ts
export type PaginatedRepositoryPickerRepos$data = {
  readonly enterpriseRepositories: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly " $fragmentSpreads": FragmentRefs<"RepositoryPickerFragment">;
      } | null | undefined;
    } | null | undefined> | null | undefined;
    readonly totalCount: number;
  };
  readonly id: string;
  readonly " $fragmentType": "PaginatedRepositoryPickerRepos";
};